+++ blm_prometheus/pkg/log/logger.go
@@ -1,3 +1,18 @@
+/*
+ * Copyright (c) 2019 TAOS Data, Inc. <jhtao@taosdata.com>
+ *
+ * This program is free software: you can use, redistribute, and/or modify
+ * it under the terms of the GNU Affero General Public License, version 3
+ * or later ("AGPL"), as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package log
 
 import (
@@ -9,10 +24,10 @@ import (
 
 const (
 	flag       = log.Ldate | log.Ltime | log.Lshortfile
-	preDebug   = "[DEBUG]"
-	preInfo    = "[INFO]"
-	preWarning = "[WARNING]"
-	preError   = "[ERROR]"
+	preDebug   = "[DEBUG] "
+	preInfo    = "[INFO] "
+	preWarning = "[WARNING] "
+	preError   = "[ERROR] "
 )
 
 var (
@@ -22,13 +37,6 @@ var (
 	warningLogger *log.Logger
 	errorLogger   *log.Logger
 )
-
-//func GetInstance(fileName string) *log.Logger {
-//	once.Do(func() {
-//		logger = Init(fileName)
-//	})
-//	return logger
-//}
 
 func Init(fileName string) {
 	logFile, err := os.OpenFile(fileName, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0644)
